Create Reservations component
